+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [UNRELEASED] 
 * [#488] Updated missing proposer address in proposals.
+* [#486] Fixed delegators voting power display on gov proposals
 * [#493] Fixed validators address showing as undefined in links.
+
 ## v0.41.x-13
 * [#484] Replaced delegation list displayed under validator page with total number of delegations. 
 * [#509] Added display list of addresses available for user to login with when connecting the Ledger.